+++ fabric/src/main/java/dev/tapetum/shaders/mixin/MixinTerrainVertexBuffer.java
@@ -2,9 +2,7 @@ package dev.tapetum.shaders.mixin;
 
 import com.mojang.blaze3d.vertex.VertexBuffer;
 import com.mojang.blaze3d.vertex.VertexFormat;
-import com.mojang.math.Matrix4f;
 import dev.tapetum.shaders.TapetumShaders;
-import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
@@ -13,13 +11,18 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 @Mixin(VertexBuffer.class)
 public abstract class MixinTerrainVertexBuffer {
-    @Shadow private int id;
-    @Shadow private int vertexCount;
-    @Shadow @Final private VertexFormat format;
+    @Shadow private int vertextBufferId;
+    @Shadow private int indexCount;
+    @Shadow private VertexFormat.IndexType indexType;
+    @Shadow private VertexFormat.Mode mode;
+    @Shadow private VertexFormat format;
 
-    @Inject(method = "draw", at = @At("HEAD"), cancellable = true)
-    private void tapetum$drawTerrain(Matrix4f modelView, int mode, CallbackInfo ci) {
-        if (TapetumShaders.getPipelineManager().drawTerrain(id, vertexCount, format, modelView, mode)) {
+    @Inject(method = "drawChunkLayer", at = @At(value = "INVOKE",
+        target = "Lcom/mojang/blaze3d/systems/RenderSystem;drawElements(III)V"), cancellable = true)
+    private void tapetum$drawTerrain(CallbackInfo ci) {
+        int indices = org.lwjgl.opengl.GL11.glGetInteger(org.lwjgl.opengl.GL15.GL_ELEMENT_ARRAY_BUFFER_BINDING);
+        if (TapetumShaders.getPipelineManager().drawTerrain(vertextBufferId, indices, indexCount,
+                indexType.asGLType, format, mode)) {
             ci.cancel();
         }
     }